+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - DocHandlerFacade.findHandler() was not looking for all registered handlers in some case.
 - Fixed missing attributes 'map-all' in complex step xsd.
+- Markdown handler ext table with extra spaces on header
 
 ## [3.0.9] - 2023-09-24